+++ internal/consts/consts.go
@@ -53,6 +53,8 @@ const (
 	FieldLeaseRenewable                = "lease_renewable"
 	FieldDepth                         = "depth"
 	FieldDataJSON                      = "data_json"
+	FieldKey                           = "key"
+	FieldValue                         = "value"
 	FieldDN                            = "dn"
 	FieldRole                          = "role"
 	FieldRoles                         = "roles"